Все работает кроме редактирования

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -1,12 +1,14 @@
 package ru.kata.spring.boot_security.demo.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.service.RoleService;
 import ru.kata.spring.boot_security.demo.service.UserService;
+
 import java.security.Principal;
 
 @Controller
@@ -24,48 +26,42 @@ public class AdminController {
     }
 
     @GetMapping
-    public String getAllUsers(ModelMap model, Principal principal) {
-        System.out.println(principal.getName());
-        model.addAttribute("allUser", userService.getAllUsers());
-        //model.addAttribute("admin", userService.getUserByUsername(principal.getName()));
-        //model.addAttribute("roles", roleService.getAllRoles());
-        return "view-users";
-    }
-
-    @GetMapping(value = "/add_user")
-    public String addUser(ModelMap model) {
-        model.addAttribute("user", new User());
-        model.addAttribute("add", true);
+    public String getAllUsers(ModelMap model, @ModelAttribute("newUser") User newUser
+            , Principal principal) {
+        User authenticatedUser = userService.getUserByUsername(principal.getName());
+        model.addAttribute("authenticatedUser", authenticatedUser);
+        model.addAttribute("authenticatedUserRoles", authenticatedUser.getRoles());
+        model.addAttribute("users", userService.getAllUsers());
         model.addAttribute("roles", roleService.getAllRoles());
-        return "add_edit";
+        return "admin-page";
     }
 
+
+    // ADD USER
     @PostMapping
-    public String saveUser(@ModelAttribute("user") User user, ModelMap model) {
-        userService.saveUser(user);
-        model.addAttribute("add", true);
-
+    public String saveUser(@ModelAttribute("user") User newUser) {
+        userService.saveUser(newUser);
         return "redirect:/admin";
     }
 
-    @GetMapping(value = "/edit_user/{id}")
-    public String editUser(ModelMap model, @PathVariable("id") Long id) {
-        model.addAttribute("user", userService.getUserById(id));
-        model.addAttribute("add", false);
-        model.addAttribute("roles", roleService.getAllRoles());
-        return "add_edit";
-    }
-
-    @PatchMapping("/edit_user/{id}")
-    public String updateUser(@ModelAttribute("user") User user, ModelMap model) {
-        userService.editUser(user);
-        model.addAttribute("add", false);
+    // EDIT USER
+    @PatchMapping("/edit")
+    public String updateUser(@ModelAttribute("editedUser") User editedUser) {
+        userService.editUser(editedUser);
         return "redirect:/admin";
     }
 
+    // DELETE USER
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable("id") Long id) {
-        userService.deleteUser(id);
-        return "redirect:/admin";
+    public String deleteUser(@PathVariable("id") Long id, Principal principal) {
+        boolean checkDeletingActivedUser = userService.getUserByUsername(principal.getName()).equals(userService.getUserById(id));
+
+        if (checkDeletingActivedUser) {
+            userService.deleteUser(id);
+            return "redirect:/process_login";
+        } else {
+            return "redirect:/admin";
+        }
+
     }
 }
